+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed Bun build and output.
+- Fixed Deno build and output.
 
 ## [0.3.0] - 2025-12-24